move vg_lite_tvg dir dev -> others
