+++ Kconfig
@@ -1265,6 +1265,35 @@ menu "LVGL configuration"
 			default y
 			help
 				This can save some memory, but not much. After the quick access bar is created, it can be hidden by clicking the button at the top left corner of the browsing area, which is very useful for small screen devices.
+	
+		menuconfig LV_USE_VG_LITE_THORVG
+			bool "VG-Lite Simulator"
+			default n
+			depends on LV_USE_THORVG
+			help
+				Use thorvg to simulate VG-Lite hardware behavior, it's useful
+				for debugging and testing on PC simulator. Enable LV_USE_THORVG,
+				Either internal ThorVG or external ThorVG library is required.
+
+		config LV_VG_LITE_THORVG_LVGL_BLEND_SUPPORT
+			bool "Enable LVGL blend mode support"
+			default n
+			depends on LV_USE_VG_LITE_THORVG
+
+		config LV_VG_LITE_THORVG_YUV_SUPPORT
+			bool "Enable YUV color format support"
+			default n
+			depends on LV_USE_VG_LITE_THORVG
+
+		config LV_VG_LITE_THORVG_16PIXELS_ALIGN
+			bool "Enable 16 pixels alignment"
+			default y
+			depends on LV_USE_VG_LITE_THORVG
+
+		config LV_VG_LITE_THORVG_THREAD_RENDER
+			bool "Enable multi-thread render"
+			default n
+			depends on LV_USE_VG_LITE_THORVG
 	endmenu
 
 	menu "Devices"
@@ -1444,35 +1473,6 @@ menu "LVGL configuration"
 		config LV_USE_EVDEV
 			bool "Use evdev input driver"
 			default n
-
-		menuconfig LV_USE_VG_LITE_THORVG
-			bool "VG-Lite API Simultor"
-			default n
-			depends on LV_USE_THORVG
-			help
-				Use thorvg to simulate VG-Lite hardware behavior, it's useful
-				for debugging and testing on PC simulator. Enable LV_USE_THORVG,
-				Either internal ThorVG or external ThorVG library is required.
-
-		config LV_VG_LITE_THORVG_LVGL_BLEND_SUPPORT
-			bool "Enable LVGL blend mode support"
-			default n
-			depends on LV_USE_VG_LITE_THORVG
-
-		config LV_VG_LITE_THORVG_YUV_SUPPORT
-			bool "Enable YUV color format support"
-			default n
-			depends on LV_USE_VG_LITE_THORVG
-
-		config LV_VG_LITE_THORVG_16PIXELS_ALIGN
-			bool "Enable 16 pixels alignment"
-			default y
-			depends on LV_USE_VG_LITE_THORVG
-
-		config LV_VG_LITE_THORVG_THREAD_RENDER
-			bool "Enable multi-thread render"
-			default n
-			depends on LV_USE_VG_LITE_THORVG
 	endmenu
 
 	menu "Examples"
