+++ scripts/code-format.cfg
@@ -30,7 +30,6 @@
 --exclude=../src/lv_conf_internal.h
 --exclude=../src/core/lv_obj_style_gen.c
 --exclude=../src/core/lv_obj_style_gen.h
---exclude=../src/dev/vg_lite_tvg/vg_lite.h
 --exclude=../src/libs/gif/gifdec.c
 --exclude=../src/libs/gif/gifdec.h
 --exclude=../src/libs/lodepng/lodepng.c
@@ -42,6 +41,7 @@
 --exclude=../src/libs/tjpgd/tjpgdcnf.h
 --exclude=../src/libs/thorvg
 --exclude=../src/libs/lz4
+--exclude=../src/others/vg_lite_tvg/vg_lite.h
 --exclude=../tests/unity/unity.c
 --exclude=../tests/unity/unity_internals.h
 --exclude=../tests/unity/unity_support.c
