+++ src/draw/vg_lite/lv_draw_vg_lite_type.h
@@ -21,7 +21,7 @@ extern "C" {
 #include "../lv_draw.h"
 
 #if LV_USE_VG_LITE_THORVG
-#include "../../dev/vg_lite_tvg/vg_lite.h"
+#include "../../others/vg_lite_tvg/vg_lite.h"
 #else
 #include <vg_lite.h>
 #endif
