+++ src/draw/vg_lite/lv_vg_lite_utils.h
@@ -20,7 +20,7 @@ extern "C" {
 
 #include <stdbool.h>
 #if LV_USE_VG_LITE_THORVG
-#include "../../dev/vg_lite_tvg/vg_lite.h"
+#include "../../others/vg_lite_tvg/vg_lite.h"
 #else
 #include <vg_lite.h>
 #endif
